minors

+++ src/homelab/__main__.py
@@ -15,6 +15,8 @@ from homelab.logger import log
 
 
 def create_symlink(source_path: Path, target_path: Path, decision: Optional[str] = None) -> str | None:
+    force: bool = decision is not None
+
     log.info(f"Config path is: {source_path}")
 
     if not source_path.exists():
@@ -30,54 +32,37 @@ def create_symlink(source_path: Path, target_path: Path, decision: Optional[str]
         log.info(f"Target path {target_path} already exists.")
 
         if target_path.resolve() == source_path:
-            log.info(f"Path {target_path} is a symbolic link to the source. Skipping symbolic link creation.")
+            print(f"Path {target_path} is a symbolic link to the source. Skipping symbolic link creation.")
             return decision
 
         if decision is None:
-            log.info("How do you want me to resolve the conflict? (rename/delete/skip) ")
-            answer = input().lower().strip()
-            if answer.startswith("rename"):
+            print("How do you want me to resolve the conflict? (rename/delete/skip (!)) ")
+            match input().lower().strip().split("!"):
+                case [decision]:
+                    force = False
+                case [decision, _]:
+                    force = True
+
+        match decision:
+            case "rename":
                 new_target_path = target_path.with_name(f"{target_path.name}_bak_{int(time.time())}")
-                log.info(f"Renaming the existing target to {new_target_path}")
+                print(f"Renaming the existing target to {new_target_path}")
                 shutil.move(str(target_path), str(new_target_path))
-                decision = "r"
-            elif answer.startswith("delete"):
-                log.info(f"Deleting the existing target {target_path}")
-                if target_path.is_dir():
-                    shutil.rmtree(str(target_path))
-                else:
-                    target_path.unlink()
-                decision = "d"
-            elif answer.startswith("skip"):
-                decision = "s"
+            case "delete":
+                print(f"Deleting the existing target {target_path}")
+                shutil.rmtree(str(target_path)) if target_path.is_dir() else target_path.unlink()
+            case "skip":
                 log.info("Skipping symbolic link creation.")
-            else:
-                log.info("Invalid answer. Please answer rename, delete or skip.")
-                return decision
 
-            log.info("Do you want to apply this action to all future conflicts? (yes/no) ")
-            answer = input().lower()
-            if answer.startswith("yes"):
-                log.info(f"Will use {decision} for all future conflicts.")
-            else:
-                decision = None
-        elif decision == "r":
-            new_target_path = target_path.with_name(f"{target_path.name}_bak_{int(time.time())}")
-            log.info(f"Renaming the existing target to {new_target_path}")
-            shutil.move(str(target_path), str(new_target_path))
-        elif decision == "d":
-            log.info(f"Deleting the existing target {target_path}")
-            if target_path.is_dir():
-                shutil.rmtree(str(target_path))
-            else:
-                target_path.unlink()
-        elif decision == "s":
-            log.info("Skipping symbolic link creation.")
-            return decision
+    if decision != "skip":
+        log.info("Creating symbolic link...")
+        target_path.symlink_to(source_path)
+        log.info("Symbolic link created successfully!")
 
-    log.info("Creating symbolic link...")
-    target_path.symlink_to(source_path)
-    log.info("Symbolic link created successfully!")
+    if force:
+        print(f"Will use {decision} for all future conflicts.")
+    else:
+        decision = None
 
     return decision
